binaries

+++ bin/cleanup-envs.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 PROJECT="_null_"
-PROJECTS="testing development test bluegreen abdeployment mlbparks"
+PROJECTS="`oc get projects -o name | sed -e 's/project\///'`"
 DEMO_USER="demo"
 
 . ../libs/functions
@@ -11,6 +11,16 @@ usage() {
   echo "${0} -p [ project_name | all ]"
   exit
 }
+
+# No point doing anything if we don't have any projects!
+if [ `echo ${PROJECTS} | wc -w` -eq 0 ]
+then
+  echo "No projects to delete here!"
+  exit
+fi
+
+#echo "Existing projects:"
+#echo "${PROJECTS}"
 
 # parse arguments
 while getopts ":p:" opt
@@ -37,23 +47,33 @@ do
   esac
 done
 
-# bail out if we don't have the right arguments
+# ask for project if none passed as args and we have something to delete!
 if [ "${PROJECT}" = "_null_" ]
 then
-  usage
+  echo "Choose project to delete:"
+  select PROJECT in ${PROJECTS}
+  do
+    break
+  done
 fi
 
 # START
 echo
-echo "REMOVE DEMO APPS..."
+echo "REMOVE PROJECT(S)..."
 echo
 
+# Need to login into demo namespace...
 echo "Logging in as ${DEMO_USER} user..."
 oc login -u ${DEMO_USER}
 
-for p in ${PROJECT}
-do
-  echo "Deleting project \"${p}\"..."
-  oc delete project ${p} 2>/dev/null
-done
-
+# Finally, do the dirty work
+if [ "${PROJECT}" != "_null_" -a "${PROJECT}" != "" ]
+then
+  for p in ${PROJECT}
+  do
+    echo "Deleting project \"${p}\"..."
+    oc delete project ${p} 2>/dev/null
+  done
+else
+  exit
+fi
